requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator==2.1.1
pydantic==2.7.1
python-multipart==0.0.9
bcrypt==4.1.2
passlib==1.7.4